chore(ts): convert FieldToken to TypeScript
`src/FieldToken/index.jsx` -> `src/FieldToken/index.tsx`. Unblocks
`TokenInput` and `MultiSelect`.

Consumer impact: none. `FieldToken` is still a `declare const` stub in
`src/index.ts`.

Typed as `React.forwardRef<HTMLDivElement, FieldTokenProps>`, preserving the
named inner function so the displayName is unchanged. `onDismiss` is typed
`(label: string) => void`, matching how it is invoked; both existing call
sites remain assignable (`MultiSelect` passes a zero-arg callback).

Props are deliberately typed strictly, with no `[x: string]: unknown` index
signature, because `FieldToken` genuinely does not spread rest props onto its
root node.

That is worth flagging for the future MultiSelect conversion.
`MultiSelect/index.js:78` does:

    <FieldToken
      label={tokenLabel}
      onDismiss={() => removeSelectedItem(item)}
      disabled={disabled}
      {...getSelectedItemProps({ selectedItem: item, i })}
    />

Since `FieldToken` only destructures `label`, `onDismiss`, `disabled` and
`testId`, everything downshift returns from `getSelectedItemProps` is
silently dropped -- except `ref`, which React applies specially. That likely
means downshift's selected-item keyboard navigation is not wired up.

Adding an index signature here would paper over that. Leaving the type honest
means converting MultiSelect will surface it as a compile error, which is the
right time to decide whether to spread the props or drop them intentionally.
Not fixing it here: it is a MultiSelect behavior change, not a conversion.

`propTypes` retained per convention. Verified: `tsc --noEmit`, `lint`,
`build` clean; 303 tests pass.
